api/api/requirements.txt
django
djangorestframework
Pillow
django-filter
psycopg2
python-dotenv
django-cors-headers
django-storages[azure]
